Added collect code, removed unnecessary script files
