+++ Astral/Assets/Scripts/Collectables/Collect.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8b05bf13d8def684ebd798f841ebc1d6
+guid: 6446de61fd05e0548b247bfd2b0f1a30
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
