--- Astral/Assets/Scripts/Collectables/Collection.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 35579e9184b36c74e9c76ce5831d00dc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
