--- Astral/Assets/Scripts/Collectables/CollectionSaveSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 45d13b93fa14e6b4fa0a2d6b11e03cac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
