--- Astral/Assets/Scripts/Collectables/CollectionUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 15c29368dad130d40bd2405261cc164c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
