--- Astral/Assets/Scripts/Collectables/Item.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c81e842f3ff7d1a408a0f844b8ef8c84
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
