--- Astral/Assets/Scripts/Collectables/NewPickup.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dd3de4def52d3754ebc65d3a9384c79e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
