--- Astral/Assets/Scripts/Collectables/Pickup.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 199336243a427bf4782f4734a54ac0e8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
