+++ Astral/Assets/material/collectableTemp.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7b202199c48143946bcf1e30cacb6b7f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
