+++ Astral/Assets/prefabs/Collectable.prefab.meta
